# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Enterprise Platform API - LMS, Tasks, Files, Automations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, manager, employee]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /modules/{moduleId}/lessons:
    get:
      operationId: getModuleLessons
      tags: [courses]
      summary: Get lessons for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [courses]
      summary: Create a lesson
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [courses]
      summary: Mark a lesson as complete
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              required: [userId]
      responses:
        "200":
          description: Lesson completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, completed]
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks/{id}/comments:
    get:
      operationId: getTaskComments
      tags: [tasks]
      summary: Get comments for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addTaskComment
      tags: [tasks]
      summary: Add comment to task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                content:
                  type: string
              required: [userId, content]
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        department:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, isActive, createdAt]
...
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        department:
          type: string
          nullable: true
      required: [email, name, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        department:
          type: string
          nullable: true
        isActive:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, moduleId, title, order, createdAt]
    Progress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        lessonId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [not_started, in_progress, completed]
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, userId, courseId, status, createdAt]
...
    CreateLessonRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        order:
          type: integer
      required: [title, order]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: integer
          nullable: true
        createdById:
          type: integer
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        assignee:
          $ref: "#/components/schemas/User"
          nullable: true
        createdBy:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, title, status, priority, createdById, createdAt, updatedAt]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: integer
          nullable: true
        createdById:
          type: integer
        dueDate:
          type: string
          format: date-time
          nullable: true
      required: [title, status, priority, createdById]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
    Comment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, taskId, userId, content, createdAt]
...