# lib/api-zod/src/generated/types/createLessonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Platform API - LMS, Tasks, Files, Automations
 * OpenAPI spec version: 1.0.0
 */

export interface CreateLessonRequest {
  title: string;
  content?: string | null;
  videoUrl?: string | null;
  duration?: number | null;
  order: number;
}